+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - New development environment with webpack and babel to use ES6
 - New shortcuts for linux
+- Added price lines in data object (popup view).
 
 ### Changed
 - Shortcuts behaviour. Show only transports, packages o all.